requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-countries==7.2.1
django-crispy-forms==1.12.0
django-environ==0.4.5
django-summernote==0.8.11.6
django-taggit==1.4.0
django-tinymce==3.3.0
idna==2.10
oauthlib==3.1.1
Pillow==8.2.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
